// src/app/shared/models/getter.ts
export class Getter {
  getterId?: number;
  getterName?: string;
  providerId?: number;
  providerName?: string;
  providerPhoneNumber?: string;
  getterTransactionDate?: Date;
  getterAmt?: number;
  getterStatus?: string;
  getterDate?: Date;
  getterDateString?: string;
  getterTDateString?: string;
}